Fix client (scope) model test placement
Fixes: #17212
