+++ testsuite/model/src/test/java/org/keycloak/testsuite/model/user/UserModelTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.testsuite.model;
+package org.keycloak.testsuite.model.user;
 
 import org.keycloak.component.ComponentModel;
 import org.keycloak.models.Constants;
@@ -32,6 +32,8 @@ import org.keycloak.storage.UserStorageProviderFactory;
 import org.keycloak.storage.UserStorageProviderModel;
 import org.keycloak.storage.UserStorageUtil;
 import org.keycloak.storage.user.UserRegistrationProvider;
+import org.keycloak.testsuite.model.KeycloakModelTest;
+import org.keycloak.testsuite.model.RequireProvider;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
